refactor app routes
