+++ src/routes/CharacterDetail/CharacterDetail.tsx
