+++ src/routes/CharacterDetail/index.ts
